+++ src/app/community/page.tsx
@@ -1,8 +1,9 @@
 import Cards from "@/Components/Community/Cards";
+import Filter from "@/Components/Community/Filter";
 import React from "react";
 
 const ComuunityPage = () => {
-    return <Cards />;
+    return <Filter />;
 };
 
 export default ComuunityPage;
